Python importing sucks + reworked constraint processor
Constraint processor has been reworked
 - less functions
 - split up amongst multiple files

So for whatever reason I was unable to import from within the
processor/ subdir, so all the folder organization has been thrown out
