--- src/optimization/constraint_builder/src/processor/constraintprocesser.py
@@ -1,563 +1,0 @@
-'''
-Constraint Processor
-
-This file does the actual processing:
- - generating constraint classes
- - reading files
- - saving to custom format
-
-
-Michael Gorbunov
-NJDEP / NJFS
-Started 05/21/2022
-'''
-
-# import optimization.constraint_builder.varname_dataclasses as models
-import processor.models as models
-from pathlib import Path
-from typing import List, Dict
-import itertools
-import csv
-import sys
-import re
-
-
-# TODO:
-# [ ] Remove main method, this file is not meant to actually be run
-# [~] Come up with better names (groups, vars, varGroupList, etc is _really_ confusing)
-# [ ] Be more strict with delimiting
-#    - Only allow for one character in the delimiting ?
-# [~] Have dataclass for constraint classes
-#    - There should be a sense of an abstract constraint class (group members)
-#      and a concrete constraint class (which can be exactly compiled into constraints)
-# [ ] Use Set instead of List for tag_groups in class models.VarTagsInfo
-# [ ] Adjust architecture so all method for compiling constraints groups -> constraints
-#      live inside varname_dataclasses.py
-# [ ] Instead of returning lists of copmiled constraints, return an iterator or generator function
-# [ ] Have a way to be Exclusive or Inclusive with categories ??
-#    - We may want a constriant that for example applies to all but 2 tree species, and this would allow
-#      for generalizing the scripts for other obj inputs (other states ?)
-# [ ] Add type checks for important processing functions
-
-def main():
-	# Input
-	objCSVPath = getCSVFilepath("Objective file: ")
-	varNamesRaw = readAllObjVarnames(objCSVPath)
-	delim = input("Delimiter: ")
-
-	# Linting
-	errMsg = lintVarNames(varNamesRaw, delim)
-	if errMsg != None:
-		errorAndExit(errMsg)
-
-	# Processing
-	varnameTags = splitVarsToTags(varNamesRaw, delim)
-	tagGroupMembersList = makeTagGroupMembersList(varnameTags)
-
-	# Input
-	tagGroupNames = getTagGroupNames(tagGroupMembersList)
-
-	# Processing
-	varTagGroupsData = makeVarTagsInfoObject(varNamesRaw, delim, tagGroupNames)
-
-	constrAllPLSQBySpecies = models.StandardConstraintGroup(
-		selected_tags = {"species": ["167N", "167S", "409"], "year": ["2021", "2025", "2030"], "mng": ["PLSQ"]},
-		split_by_groups = ["species"],
-		constr_prefix = "PLSQ_ALL"
-	)
-
-	constrAllBySpeciesByYear = models.StandardConstraintGroup(
-		selected_tags = {"species": ["167N", "409"], "year": ["2021", "2025", "2030", "2050"], "mng": ["PLSQ", "PLWF", "RBWF", "STQO", "TB", "TBWF", "RxB"]},
-		split_by_groups = ["species", "year"],
-		constr_prefix = "All"
-	)
-
-	print("\n" * 3)
-	constList = compileStandardConstraintGroup(varTagGroupsData, constrAllPLSQBySpecies)
-	for const in constList:
-		print(f"{const.name} \t|\t{' + '.join(['_'.join(x) for x in const.var_tags])}   \t{str(const.compare_type) + str(const.compare_value)}")
-
-	print("\n" * 3)
-	print("[[ Now generating constraints split by species & year ]]")
-	constList = compileStandardConstraintGroup(varTagGroupsData, constrAllBySpeciesByYear)
-	for const in constList:
-		print(f"{const.name} \t|\t{' + '.join(['_'.join(x) for x in const.var_tags])}   \t{str(const.compare_type) + str(const.compare_value)}")
-
-	
-
-
-
-
-
-#
-# Processing
-#
-
-def compileStandardConstraintGroup (varInfo: models.VarTagsInfo, stdConGroup: models.StandardConstraintGroup) -> List[models.CompiledConstraint]:
-	conDict = {}
-
-	#	
-	# Generate names of all constraints
-	conNames = []
-	conNameTags = []
-
-	# Iterating through varInfo.tag_order guarantees the order is correct
-	for tagGroup in varInfo.tag_order:
-		if tagGroup in stdConGroup.split_by_groups:
-			conNameTags.append(stdConGroup.selected_tags[tagGroup])
-	
-	for conTags in itertools.product(*conNameTags):
-		conNames.append("_".join([stdConGroup.constr_prefix] + list(conTags)))
-	if len(conNames) == 0:
-		conNames = [stdConGroup.constr_prefix]
-	
-	# The names are used as keys in the constraint dictionary, so we
-	# populate the dictionary with empty lists
-	for name in conNames:
-		conDict[name] = []
-	
-	# print(conDict)
-	
-
-	#
-	# Generate the indicies that get split by
-	tagListsByGroup = [] # this is a 2D array
-	for tagGroup in varInfo.tag_order:
-		tagListsByGroup.append(stdConGroup.selected_tags[tagGroup])
-	
-	# TODO: Refactor to avoid needing indToSplitBy
-	indToSplitBy = []
-	for splitByGroup in stdConGroup.split_by_groups:
-		ind = varInfo.tag_order.index(splitByGroup)
-		indToSplitBy.append(ind)
-	indToSplitBy.sort()
-
-
-	#
-	# Generate all potential variables and split into multiple constraints
-	for varTags in itertools.product(*tagListsByGroup):
-		varTags = list(varTags)
-		if not (varTags in varInfo.all_vars):
-			continue
-
-		tagsToSplitBy = []
-		for ind in indToSplitBy:
-			tagsToSplitBy.append(varTags[ind])
-		conName = "_".join([stdConGroup.constr_prefix] + tagsToSplitBy)
-
-		conDict[conName].append(varTags)
-
-	# Check for potential empty constraints & convert to list
-	allKeys = list(conDict.keys())
-	compiledConList = []
-
-	for key in allKeys:
-		numVars = len(conDict[key])
-
-		if numVars == 0:
-			printInfo(f"Found empty constraint {key}")
-		else:
-			compiledConList.append(
-				models.CompiledConstraint(
-					name = key,
-					var_tags = conDict[key],
-					var_coeffs = [stdConGroup.default_coef] * numVars,
-					compare_type = stdConGroup.default_compare,
-					compare_value = stdConGroup.default_rightside
-				)
-			)
-	
-	return compiledConList
-
-
-
-def makeVarTagsInfoObject (varNamesRaw: List[str], delim: str, tagGroupNames: List[str]) -> models.VarTagsInfo:
-	'''
-		DOES NOT LINT. Make sure to call the linting functions before passing into this
-	'''
-	varnameTags = splitVarsToTags(varNamesRaw, delim)
-	tagGroupMembersList = makeTagGroupMembersList(varnameTags)
-
-	tagGroupsDict = {}
-	for ind, name in enumerate(tagGroupNames):
-		tagGroupsDict[name] = tagGroupMembersList[ind]
-
-	return models.VarTagsInfo(
-		tag_order = tagGroupNames,
-		all_vars = varnameTags,
-		tag_groups = tagGroupsDict
-	)
-
-
-def makeVarTagsInfoObjectFromFile (filestr: str, delim: str, tagGroups: List[str]) -> models.VarTagsInfo:
-	'''
-		DOES NOT LINT. It will silently return None
-	'''
-	if filestr == None or not Path(filestr).is_file():
-		return None
-
-	varNamesRaw = readAllObjVarnames(Path(filestr))
-	return makeVarTagsInfoObject(varNamesRaw, delim, tagGroups)
-
-
-
-
-
-
-def makeTagGroupMembersList (allVarTags: List[List[str]]) -> List[List[str]]:
-	'''
-		Takes a list of variables in the form
-
-		[	[167S, 2021, NM],
-		 	[167S, 2025, NM],
-		 	[167N, 2025, NM],
-		 	[167N, 2030, SBP]	]
-
-		and returns lists of each individual tag group,
-		in the order they appear within the variables.
-
-		[	[167S, 167N],
-			[2021, 2025, 2030],
-			[NM, SBP]	] 
-	'''
-	tagGroupMembers = []
-	for i in range(len(allVarTags[0])):
-		tagGroupMembers.append(set())
-	
-	for varTags in allVarTags:
-		for ind, groupID in enumerate(varTags):
-			tagGroupMembers[ind].add(groupID)
-
-	for ind, tags in enumerate(tagGroupMembers):
-		tagGroupMembers[ind] = list(tags)
-		tagGroupMembers[ind].sort()
-
-	return tagGroupMembers
-
-
-def splitVarsToTags (rawVarnames: List[str], delim: str) -> List[List[str]]:
-	'''
-	Splits variables up by their delimiter
-	input:  [ "167S_2021_NM",          "167N_2030_SBP" ] 
-	output: [ ["167S", "2021", "NM"], ["167N", "2030", "SBP"] ]
-	'''
-	return [splitRawVarIntoTags(x, delim) for x in rawVarnames]
-
-
-def getNumTagsInRawVar (varnameRaw: str, delim: str) -> int:
-	return len(splitRawVarIntoTags(varnameRaw, delim))
-
-
-def splitRawVarIntoTags (varnameRaw, delim: str) -> List[str]:
-	return re.split(delim + "+", varnameRaw)
-
-
-
-
-
-#
-# Linting
-#
-
-def lintVarNames (varNamesRaw: List[str], delim: str) -> str:
-	'''
-	Goes through the list of variable names and checks that they're nice
-	 - returns None if there are no erors
-	 - returns an error message string if there are errors
-	'''
-	VALID_DELIMITERS = "_-= "
-	# All group names must be exclusively alphanumeric (A-Z, a-z, 0-9)
-	TAG_NAME_REGEX = "^[A-Za-z0-9]+$"
-
-
-	# [[ Check ]] At least one variable
-	if (varNamesRaw == None or len(varNamesRaw) == 0):
-		return "Empty or None list passed in"
-
-	firstVar = varNamesRaw[0]
-
-	
-	# [[ Check ]] Delimiter is valid
-	if len(delim) == 0:
-		return f'Empty delimiter is not allowed'
-	if len(delim) > 1:
-		return f'Delimiter must be a single character. If you have variables of the form "167S--2021---SPBQ", usiing delimiter "-" is valid.'
-	if not (delim in VALID_DELIMITERS):
-		return f'Delimiter "{delim}" is invalid. It must be one of {",".join(VALID_DELIMITERS)}'
-
-
-	# [[ Check ]] Delimiter is inside of variable
-	if not delim in firstVar:
-		return f'Delimiter "{delim}" not found inside variable {firstVar}'
-
-
-	# [[ Check ]] All group names are valid
-	for var in varNamesRaw:
-		for tag in splitRawVarIntoTags(var, delim):
-			if re.search(TAG_NAME_REGEX, tag) == None:
-				return f'Found invalid groupname "{tag}" for variable "{var}". Groups must contain only letters and numbers'
-
-
-	# [[ Check ]] All variables have the same number of groups
-	numGroups = getNumTagsInRawVar(firstVar, delim)
-
-	for var in varNamesRaw[1:]:
-		testNumGroups = getNumTagsInRawVar(var, delim)
-		if testNumGroups != numGroups:
-			return f'Variable "{var}" has a different number of groups ({testNumGroups}) compare to "{firstVar}" ({numGroups})'
-
-
-def lintTagGroupName (tagName: str) -> str:
-	'''
-	Checks that the provided group name is valid, returning None if
-	there are no errors or the actual error message if there is one
-	'''
-	TAGNAME_MAX_LENGTH = 15
-	TAGNAME_MIN_LENGTH = 3
-	TAGNAME_REGEX = "^[A-Za-z0-9_-]+$"
-
-	# [[ Check ]] Size
-	if len(tagName) < TAGNAME_MIN_LENGTH:
-		return f'Groupname "{tagName}" too short. Must be at least {TAGNAME_MIN_LENGTH} characters'
-	elif len(tagName) > TAGNAME_MAX_LENGTH:
-		return f'Groupname "{tagName}" is too long. Name is {len(tagName)} characters long, but max is {TAGNAME_MAX_LENGTH}'
-	
-	# [[ Check ]] Only alphanumeric names allowed
-	if re.search(TAGNAME_REGEX, tagName) == None:
-		# Lol this is so inefficient but its kinda funny
-		for c in tagName:
-			if re.search(TAGNAME_REGEX, str(c)) == None:
-				return f'Invalid groupname "{tagName}". Illegal character "{c}"'
-
-
-def lintMultipleTagGroupNames (listTagNames: List[str]) -> str:
-	# [[ Check ]] All variable named
-	for ind, name in enumerate(listTagNames):
-		if name == None or name == '':
-			return f'Not all groups are named'
-	
-	# [[ Check ]] Individually fine
-	for name in listTagNames:
-		err = lintTagGroupName(name)
-		if err:
-			return err
-
-	# [[ Check ]] Duplicate names
-	if len(set(listTagNames)) != len(listTagNames):
-		return f'Found duplicate names'
-
-
-# TODO: Have a general constraintgroup linting? Make sure coefficients != 0, etc
-
-def lintConstrGroupName (groupName: str) -> str:
-	'''
-	Checks that the provided group name is valid, returning None if
-	there are no errors or the actual error message if there is one
-	'''
-	GROUPNAME_MAX_LENGTH = 30
-	GROUPNAME_MIN_LENGTH = 3
-	GROUPNAME_REGEX = "^[A-Za-z0-9_-]+$"
-
-	# [[ Check ]] Size
-	if len(groupName) < GROUPNAME_MIN_LENGTH:
-		return f'Groupname "{groupName}" too short. Must be at least {GROUPNAME_MIN_LENGTH} characters'
-	elif len(groupName) > GROUPNAME_MAX_LENGTH:
-		return f'Groupname "{groupName}" is {len(groupName) - GROUPNAME_MAX_LENGTH} characters too long.'
-	
-	# [[ Check ]] Only alphanumeric names allowed
-	if re.search(GROUPNAME_REGEX, groupName) == None:
-		for c in groupName:
-			if re.search(GROUPNAME_REGEX, str(c)) == None:
-				return f'Invalid groupname "{groupName}". Illegal character "{c}"'
-
-
-def lintMultiplyConstrGroupNames (listGroupNames: List[str]) -> str:
-	# [[ Check ]] All groups named
-	for ind, name in enumerate(listGroupNames):
-		if name == None or name == '':
-			return f'Not all groups are named'
-	
-	# [[ Check ]] Individually fine
-	for name in listGroupNames:
-		err = lintConstrGroupName(name)
-		if err:
-			return err
-
-	# [[ Check ]] Duplicate names
-	if len(set(listGroupNames)) != len(listGroupNames):
-		duplicates = set([x for x in listGroupNames if listGroupNames.count(x) > 1])
-		return f'Found duplicate group names {duplicates}'
-
-
-
-
-
-#
-#  File I/O
-#
-
-def readAllObjVarnames (objCSVPath: Path) -> List[str]:
-	allVarnames = []
-
-	with open(objCSVPath, 'r') as objFile:
-		objCSVReader = csv.reader(objFile)
-		lineCount = 0
-
-		for row in objCSVReader:
-			row = [str(x).strip() for x in row]
-			lineCount += 1
-
-			if lineCount == 1:
-				continue
-
-			allVarnames.append(row[0])
-	
-	return allVarnames
-
-
-
-
-
-
-
-#
-# Command line Input
-#
-
-GROUPNAME_MAX_LENGTH = 15
-
-# TODO: Split the linting function to be entirely seperate & not called from here ?
-def getTagGroupNames (tagGroupsList: List[List[str]]) -> List[str]:
-	'''
-		Takes the list of different group ids and asks the user
-		to name them. For ex:
-
-		input:
-		[
-			['167N', '167S', '409'  ...], 
-			['2021', '2025', '2030' ...], 
-			['PLSQ', 'PLWF', 'RBWF' ...]
-		]
-
-		output (given by user):
-		[
-			"species",
-			"year",
-			"management"
-		]
-	'''
-	tagGroupNames = []
-
-	for tagGroup in tagGroupsList:
-		sampleMembers = None
-		if len(tagGroup) < 3:
-			sampleMembers = tagGroup
-		else:
-			sampleMembers = tagGroup[:3] + ["..."]
-		sampleStr = ", ".join(sampleMembers)
-
-		name = input(f"\nGroup members: {sampleStr}\nName for group? ").strip()
-
-		errMsg = lintTagGroupName(name)
-		if errMsg:
-			errorAndExit(errMsg)
-
-		tagGroupNames.append(name)
-	
-	return tagGroupNames
-
-
-
-
-
-def getCSVFilepath (message: str) -> Path:
-	'''
-		Takes user input for a path and makes
-		sure that the file is a csv and exists.
-
-		In the event of an error, this will terminate
-		the program. The returned object is never None.
-	'''
-	pathStr = input(message).strip()
-
- 	# Sometimes the passed in directory is surrounded by quotes
-	# This checks for & removes them
-	# ex: '/home/velcro' -> /home/velcro
-	firstAndLastChars = pathStr[0] + pathStr[-1]
-	if firstAndLastChars == "''" or firstAndLastChars == '""':
-		pathStr = pathStr[1:-1]
-
-	pathObj = Path(pathStr)
-
-	if not pathObj.exists():
-		errorAndExit("Supplied filepath does not exist")
-	if pathObj.is_dir():
-		errorAndExit("Supplied filepath is to a directory")
-
-	fileExt = pathObj.suffix
-
-	if fileExt != '.csv':
-		errorAndExit("Supplied filepath is not a csv")
-
-	return pathObj
-
-
-def errorAndExit (errMessage: str):
-	print()
-	print('\t[[ XX Error ]]')
-	print(f'{errMessage}')
-	print()
-	print("Aborting.")
-	sys.exit(1)
-
-
-def printWarning (warnMessage: str):
-	print(f'\t[[ !! Warning ]] : {warnMessage}')
-
-
-def printInfo (infoMessage: str):
-	print(f'\t[[ ~~ Info ]] : {infoMessage}')
-
-
-
-
-
-
-
-
-
-
-
-
-
-if __name__ == '__main__':
-	main()
-
-
-
-
-'''
-⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣀⣤⣤⣤⣀⣀⣀⣀⡀⠀⠀⠀⠀⠀⠀⠀
-⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣼⠟⠉⠉⠉⠉⠉⠉⠉⠙⠻⢶⣄⠀⠀⠀⠀⠀
-⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣾⡏⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠙⣷⡀⠀⠀⠀
-⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣸⡟⠀⣠⣶⠛⠛⠛⠛⠛⠛⠳⣦⡀⠀⠘⣿⡄⠀⠀
-⠀⠀⠀⠀⠀⠀⠀⠀⠀⢠⣿⠁⠀⢹⣿⣦⣀⣀⣀⣀⣀⣠⣼⡇⠀⠀⠸⣷⠀⠀
-⠀⠀⠀⠀⠀⠀⠀⠀⠀⣼⡏⠀⠀⠀⠉⠛⠿⠿⠿⠿⠛⠋⠁⠀⠀⠀⠀⣿⡄⣠
-⠀⠀⢀⣀⣀⣀⠀⠀⢠⣿⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢻⡇⠀
-⠿⠿⠟⠛⠛⠉⠀⠀⣸⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸⡇⠀
-⠀⠀⠀⠀⠀⠀⠀⠀⣿⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢸⣧⠀
-⠀⠀⠀⠀⠀⠀⠀⢸⡿⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⣿⠀
-⠀⠀⠀⠀⠀⠀⠀⣾⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣿⠀
-⠀⠀⠀⠀⠀⠀⠀⣿⠃⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣿⠀
-⠀⠀⠀⠀⠀⠀⢰⣿⠀⠀⠀⠀⣠⡶⠶⠿⠿⠿⠿⢷⣦⠀⠀⠀⠀⠀⠀⠀⣿⠀
-⠀⠀⣀⣀⣀⠀⣸⡇⠀⠀⠀⠀⣿⡀⠀⠀⠀⠀⠀⠀⣿⡇⠀⠀⠀⠀⠀⠀⣿⠀
-⣠⡿⠛⠛⠛⠛⠻⠀⠀⠀⠀⠀⢸⣇⠀⠀⠀⠀⠀⠀⣿⠇⠀⠀⠀⠀⠀⠀⣿⠀
-⢻⣇⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣼⡟⠀⠀⢀⣤⣤⣴⣿⠀⠀⠀⠀⠀⠀⠀⣿⠀
-⠈⠙⢷⣶⣦⣤⣤⣤⣴⣶⣾⠿⠛⠁⢀⣶⡟⠉⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⡟⠀
-⢷⣶⣤⣀⠉⠉⠉⠉⠉⠄⠀⠀⠀⠀⠈⣿⣆⡀⠀⠀⠀⠀⠀⠀⢀⣠⣴⡾⠃⠀
-⠀⠈⠉⠛⠿⣶⣦⣄⣀⠀⠀⠀⠀⠀⠀⠈⠛⠻⢿⣿⣾⣿⡿⠿⠟⠋⠁⠀⠀⠀
-
-
-			 Why so Sus ?
-
-'''
